@@ -3,9 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-native-scripts": "1.3.0",
-    "jest-expo": "~20.0.0",
-    "react-test-renderer": "16.0.0-alpha.12"
+    "jest-expo": "25.0.0",
+    "react-native-scripts": "1.11.1",
+    "react-test-renderer": "16.2.0"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
@@ -20,12 +20,13 @@
   },
   "dependencies": {
     "babel-plugin-transform-remove-console": "^6.9.0",
-    "expo": "^20.0.0",
-    "react": "16.0.0-alpha.12",
-    "react-native": "^0.47.0",
+    "expo": "^25.0.0",
+    "ramda": "^0.25.0",
+    "react": "16.2.0",
+    "react-native": "^0.52.0",
     "react-native-swipeout": "^2.3.3",
     "react-navigation": "1.0.0-beta.27",
-    "react-redux": "^5.0.6",
+    "react-redux": "^5.0.7",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0"
   }
